# DAOS-8931 vea: optimize free extent insertion