--- a/src/vea/vea_free.c
+++ b/src/vea/vea_free.c
@@ -146,7 +146,8 @@ merge_free_ext(struct vea_space_info *vsi, struct vea_free_extent *ext_in,
 	daos_handle_t		 btr_hdl;
 	d_iov_t			 key, key_out, val;
 	uint64_t		*off;
-	int			 rc, opc = BTR_PROBE_LE;
+	bool			 fetch_prev = true;
+	int			 rc, del_opc = BTR_PROBE_BYPASS;
 
 	if (type == VEA_TYPE_COMPOUND)
 		btr_hdl = vsi->vsi_free_btr;
@@ -159,21 +160,56 @@ merge_free_ext(struct vea_space_info *vsi, struct vea_free_extent *ext_in,
 
 	D_ASSERT(daos_handle_is_valid(btr_hdl));
 	d_iov_set(&key, &ext_in->vfe_blk_off, sizeof(ext_in->vfe_blk_off));
+	d_iov_set(&key_out, NULL, 0);
+	d_iov_set(&val, NULL, 0);
+
+	/*
+	 * Search the one to be merged, the btree trace will be set to proper position for
+	 * later potential insertion (when merge failed), so don't change btree trace!
+	 */
+	rc = dbtree_fetch(btr_hdl, BTR_PROBE_EQ, DAOS_INTENT_DEFAULT, &key, &key_out, &val);
+	if (rc == 0) {
+		D_ERROR("unexpected extent ["DF_U64", %u]\n",
+			ext_in->vfe_blk_off, ext_in->vfe_blk_cnt);
+		return -DER_INVAL;
+	} else if (rc != -DER_NONEXIST) {
+		D_ERROR("search extent with offset "DF_U64" failed. "DF_RC"\n",
+			ext_in->vfe_blk_off, DP_RC(rc));
+		return rc;
+	}
 repeat:
 	d_iov_set(&key_out, NULL, 0);
 	d_iov_set(&val, NULL, 0);
 
-	rc = dbtree_fetch(btr_hdl, opc, DAOS_INTENT_DEFAULT, &key, &key_out,
-			  &val);
-	if (rc == -DER_NONEXIST && opc == BTR_PROBE_LE) {
-		opc = BTR_PROBE_GE;
-		goto repeat;
-	}
+	if (fetch_prev) {
+		rc = dbtree_fetch_prev(btr_hdl, &key_out, &val, false);
+		if (rc == -DER_NONEXIST) {
+			fetch_prev = false;
+			goto repeat;
+		} else if (rc) {
+			D_ERROR("search prev extent failed. "DF_RC"\n", DP_RC(rc));
+			return rc;
+		}
+	} else {
+		/*
+		 * The btree trace was set to the position for inserting the searched
+		 * one. If there is an extent in current position, let's try to merge
+		 * it with the searched one; otherwise, we'd lookup next to see if any
+		 * extent can be merged.
+		 */
+		rc = dbtree_fetch_cur(btr_hdl, &key_out, &val);
+		if (rc == -DER_NONEXIST) {
+			del_opc = BTR_PROBE_EQ;
+			rc = dbtree_fetch_next(btr_hdl, &key_out, &val, false);
+		}
 
-	if (rc == -DER_NONEXIST)
-		goto done;	/* Merge done */
-	else if (rc)
-		return rc;	/* Error */
+		if (rc == -DER_NONEXIST) {
+			goto done; /* Merge done */
+		} else if (rc) {
+			D_ERROR("search next extent failed. "DF_RC"\n", DP_RC(rc));
+			return rc;
+		}
+	}
 
 	if (type == VEA_TYPE_PERSIST) {
 		entry = NULL;
@@ -189,8 +225,7 @@ repeat:
 		return rc;
 
 	/* This checks overlapping & duplicated extents as well. */
-	rc = (opc == BTR_PROBE_LE) ? ext_adjacent(ext, &merged) :
-				     ext_adjacent(&merged, ext);
+	rc = fetch_prev ? ext_adjacent(ext, &merged) : ext_adjacent(&merged, ext);
 	if (rc < 0)
 		return rc;
 
@@ -203,7 +238,7 @@ repeat:
 			return -DER_INVAL;
 		}
 
-		if (opc == BTR_PROBE_LE) {
+		if (fetch_prev) {
 			merged.vfe_blk_off = ext->vfe_blk_off;
 			merged.vfe_blk_cnt += ext->vfe_blk_cnt;
 
@@ -218,8 +253,7 @@ repeat:
 			 */
 			if (neighbor != NULL) {
 				undock_entry(vsi, entry, type);
-				rc = dbtree_delete(btr_hdl, BTR_PROBE_BYPASS,
-						   &key_out, NULL);
+				rc = dbtree_delete(btr_hdl, del_opc, &key_out, NULL);
 				if (rc) {
 					D_ERROR("Failed to delete: %d\n", rc);
 					return rc;
@@ -231,8 +265,8 @@ repeat:
 		}
 	}
 
-	if (opc == BTR_PROBE_LE) {
-		opc = BTR_PROBE_GE;
+	if (fetch_prev) {
+		fetch_prev = false;
 		goto repeat;
 	}
 done:
@@ -270,7 +304,7 @@ compound_free(struct vea_space_info *vsi, struct vea_free_extent *vfe,
 	      unsigned int flags)
 {
 	struct vea_entry	*entry, dummy;
-	d_iov_t			 key, val;
+	d_iov_t			 key, val, val_out;
 	int			 rc;
 
 	rc = merge_free_ext(vsi, vfe, VEA_TYPE_COMPOUND, flags);
@@ -287,26 +321,19 @@ compound_free(struct vea_space_info *vsi, struct vea_free_extent *vfe,
 
 	/* Add to in-memory free extent tree */
 	D_ASSERT(daos_handle_is_valid(vsi->vsi_free_btr));
-	d_iov_set(&key, &dummy.ve_ext.vfe_blk_off,
-		  sizeof(dummy.ve_ext.vfe_blk_off));
+	d_iov_set(&key, &dummy.ve_ext.vfe_blk_off, sizeof(dummy.ve_ext.vfe_blk_off));
 	d_iov_set(&val, &dummy, sizeof(dummy));
+	d_iov_set(&val_out, NULL, 0);
 
-	rc = dbtree_update(vsi->vsi_free_btr, &key, &val);
-	if (rc != 0)
+	rc = dbtree_upsert(vsi->vsi_free_btr, BTR_PROBE_BYPASS, DAOS_INTENT_UPDATE, &key,
+			   &val, &val_out);
+	if (rc != 0) {
+		D_ERROR("Insert compound extent failed. "DF_RC"\n", DP_RC(rc));
 		return rc;
+	}
 
-	/* Fetch & operate on the in-tree record from now on */
-	d_iov_set(&key, &dummy.ve_ext.vfe_blk_off,
-		  sizeof(dummy.ve_ext.vfe_blk_off));
-	d_iov_set(&val, NULL, 0);
-
-	rc = dbtree_fetch(vsi->vsi_free_btr, BTR_PROBE_EQ, DAOS_INTENT_DEFAULT,
-			  &key, NULL, &val);
-	D_ASSERT(rc != -DER_NONEXIST);
-	if (rc)
-		return rc;
-
-	entry = (struct vea_entry *)val.iov_buf;
+	D_ASSERT(val_out.iov_buf != NULL);
+	entry = (struct vea_entry *)val_out.iov_buf;
 	D_INIT_LIST_HEAD(&entry->ve_link);
 
 	rc = free_class_add(vsi, entry);
@@ -341,7 +368,9 @@ persistent_free(struct vea_space_info *vsi, struct vea_free_extent *vfe)
 	d_iov_set(&key, &dummy.vfe_blk_off, sizeof(dummy.vfe_blk_off));
 	d_iov_set(&val, &dummy, sizeof(dummy));
 
-	rc = dbtree_update(btr_hdl, &key, &val);
+	rc = dbtree_upsert(btr_hdl, BTR_PROBE_BYPASS, DAOS_INTENT_UPDATE, &key, &val, NULL);
+	if (rc)
+		D_ERROR("Insert persistent extent failed. "DF_RC"\n", DP_RC(rc));
 	return rc;
 }
 
@@ -350,7 +379,7 @@ int
 aggregated_free(struct vea_space_info *vsi, struct vea_free_extent *vfe)
 {
 	struct vea_entry	*entry, dummy;
-	d_iov_t			 key, val;
+	d_iov_t			 key, val, val_out;
 	daos_handle_t		 btr_hdl = vsi->vsi_agg_btr;
 	int			 rc;
 
@@ -370,26 +399,18 @@ aggregated_free(struct vea_space_info *vsi, struct vea_free_extent *vfe)
 
 	/* Add to in-memory aggregate free extent tree */
 	D_ASSERT(daos_handle_is_valid(btr_hdl));
-	d_iov_set(&key, &dummy.ve_ext.vfe_blk_off,
-		  sizeof(dummy.ve_ext.vfe_blk_off));
+	d_iov_set(&key, &dummy.ve_ext.vfe_blk_off, sizeof(dummy.ve_ext.vfe_blk_off));
 	d_iov_set(&val, &dummy, sizeof(dummy));
+	d_iov_set(&val_out, NULL, 0);
 
-	rc = dbtree_update(btr_hdl, &key, &val);
-	if (rc)
+	rc = dbtree_upsert(btr_hdl, BTR_PROBE_BYPASS, DAOS_INTENT_UPDATE, &key, &val, &val_out);
+	if (rc) {
+		D_ERROR("Insert aging extent failed. "DF_RC"\n", DP_RC(rc));
 		return rc;
+	}
 
-	/* Fetch & operate on the in-tree record from now on */
-	d_iov_set(&key, &dummy.ve_ext.vfe_blk_off,
-		  sizeof(dummy.ve_ext.vfe_blk_off));
-	d_iov_set(&val, NULL, 0);
-
-	rc = dbtree_fetch(btr_hdl, BTR_PROBE_EQ, DAOS_INTENT_DEFAULT, &key,
-			  NULL, &val);
-	D_ASSERT(rc != -DER_NONEXIST);
-	if (rc)
-		return rc;
-
-	entry = (struct vea_entry *)val.iov_buf;
+	D_ASSERT(val_out.iov_buf != NULL);
+	entry = (struct vea_entry *)val_out.iov_buf;
 	D_INIT_LIST_HEAD(&entry->ve_link);
 
 	/* Add to the tail of aggregate LRU list */