--- a/src/vos/vos_container.c
+++ b/src/vos/vos_container.c
@@ -68,7 +68,7 @@ cont_df_rec_free(struct btr_instance *tins, struct btr_record *rec, void *args)
 
 static int
 cont_df_rec_alloc(struct btr_instance *tins, d_iov_t *key_iov,
-		  d_iov_t *val_iov, struct btr_record *rec)
+		  d_iov_t *val_iov, struct btr_record *rec, d_iov_t *val_out)
 {
 	struct vos_pool		*pool;
 	struct cont_df_args	*args;
@@ -128,7 +128,7 @@ cont_df_rec_fetch(struct btr_instance *tins, struct btr_record *rec,
 
 static int
 cont_df_rec_update(struct btr_instance *tins, struct btr_record *rec,
-		   d_iov_t *key, d_iov_t *val)
+		   d_iov_t *key, d_iov_t *val, d_iov_t *val_out)
 {
 	D_DEBUG(DB_DF, "Record exists already. Nothing to do\n");
 	return 0;