--- a/src/vos/vos_dtx.c
+++ b/src/vos/vos_dtx.c
@@ -265,7 +265,7 @@ dtx_hkey_cmp(struct btr_instance *tins, struct btr_record *rec, void *hkey)
 
 static int
 dtx_act_ent_alloc(struct btr_instance *tins, d_iov_t *key_iov,
-		  d_iov_t *val_iov, struct btr_record *rec)
+		  d_iov_t *val_iov, struct btr_record *rec, d_iov_t *val_out)
 {
 	struct vos_dtx_act_ent	*dae = val_iov->iov_buf;
 
@@ -315,7 +315,7 @@ dtx_act_ent_fetch(struct btr_instance *tins, struct btr_record *rec,
 
 static int
 dtx_act_ent_update(struct btr_instance *tins, struct btr_record *rec,
-		   d_iov_t *key, d_iov_t *val)
+		   d_iov_t *key, d_iov_t *val, d_iov_t *val_out)
 {
 	struct vos_container	*cont = tins->ti_priv;
 	struct vos_dtx_act_ent	*dae_new = val->iov_buf;
@@ -348,7 +348,7 @@ static btr_ops_t dtx_active_btr_ops = {
 
 static int
 dtx_cmt_ent_alloc(struct btr_instance *tins, d_iov_t *key_iov,
-		  d_iov_t *val_iov, struct btr_record *rec)
+		  d_iov_t *val_iov, struct btr_record *rec, d_iov_t *val_out)
 {
 	struct vos_tls		*tls = vos_tls_get();
 	struct vos_container	*cont = tins->ti_priv;
@@ -398,7 +398,7 @@ dtx_cmt_ent_fetch(struct btr_instance *tins, struct btr_record *rec,
 
 static int
 dtx_cmt_ent_update(struct btr_instance *tins, struct btr_record *rec,
-		   d_iov_t *key, d_iov_t *val)
+		   d_iov_t *key, d_iov_t *val, d_iov_t *val_out)
 {
 	struct vos_dtx_cmt_ent	*dce_new = val->iov_buf;
 	struct vos_dtx_cmt_ent	*dce_old;
@@ -847,7 +847,7 @@ vos_dtx_commit_one(struct vos_container *cont, struct dtx_id *dti,
 
 	d_iov_set(&riov, dce, sizeof(*dce));
 	rc = dbtree_upsert(cont->vc_dtx_committed_hdl, BTR_PROBE_EQ,
-			   DAOS_INTENT_UPDATE, &kiov, &riov);
+			   DAOS_INTENT_UPDATE, &kiov, &riov, NULL);
 	if (rc != 0)
 		goto out;
 
@@ -1018,7 +1018,7 @@ vos_dtx_alloc(struct vos_dtx_blob_df *dbd, struct dtx_handle *dth)
 	d_iov_set(&kiov, &DAE_XID(dae), sizeof(DAE_XID(dae)));
 	d_iov_set(&riov, dae, sizeof(*dae));
 	rc = dbtree_upsert(cont->vc_dtx_active_hdl, BTR_PROBE_EQ,
-			   DAOS_INTENT_UPDATE, &kiov, &riov);
+			   DAOS_INTENT_UPDATE, &kiov, &riov, NULL);
 	if (rc == 0) {
 		dae->dae_start_time = crt_hlc_get();
 		d_list_add_tail(&dae->dae_link, &cont->vc_dtx_act_list);
@@ -2514,7 +2514,7 @@ vos_dtx_act_reindex(struct vos_container *cont)
 			d_iov_set(&riov, dae, sizeof(*dae));
 			rc = dbtree_upsert(cont->vc_dtx_active_hdl,
 					   BTR_PROBE_EQ, DAOS_INTENT_UPDATE,
-					   &kiov, &riov);
+					   &kiov, &riov, NULL);
 			if (rc != 0) {
 				D_FREE(dae->dae_records);
 				dtx_evict_lid(cont, dae);
@@ -2583,7 +2583,7 @@ vos_dtx_cmt_reindex(daos_handle_t coh, void *hint)
 		d_iov_set(&kiov, &DCE_XID(dce), sizeof(DCE_XID(dce)));
 		d_iov_set(&riov, dce, sizeof(*dce));
 		rc = dbtree_upsert(cont->vc_dtx_committed_hdl, BTR_PROBE_EQ,
-				   DAOS_INTENT_UPDATE, &kiov, &riov);
+				   DAOS_INTENT_UPDATE, &kiov, &riov, NULL);
 		if (rc != 0) {
 			D_FREE(dce);
 			goto out;