@NiuYawei 2021-11-19
Merge OID refactor changes